Nuke dom4j dependency
dom4j was on the chopping block for quite a while now but
managed to live so far purely because a migration to java's
XML api would cause more bugs than would be the case if the
dependency would survive as-is.

Therefore, this commit is likely to introduce a few nasty
bugs but that is the cycle of life.

The last straw why dom4j was removed is because dom4j does
not properly support multi-classloader environments such
as is the case with SLL.