Make project open-source, move to stianloader
Do note that while it is now open-source, it won't necessarily be
open-contribution. This is mainly aimed at reducing the risk of allegations
of plagiarism, but realistically the risk should be near 0 now (since
I already handed in the relevant assignment).

This commit also attaches some internal design documents used
in my "Besondere Lernleistung" - they are in german, so I am not too
sure if they will be all too useful for you - dear reader (and GH
won't render them anyways :P - you need to use PlantUML to render them,
but that can easily be done online if you REALLY need to).