Resolve version ranges by newest version
This behaviour might be a bit counterintuitive to some, completely intuitive
to others. Truth be told I haven't checked what maven thinks about this all
but I'd say that this is the most intuitive way of solving the problem for
me (as otherwise the version would be known, so why go with a version range?).
Of course the risk I'm taking with this commit is that versions may break
other versions, but in reality anyone making use of version ranges should be
very aware of this serious risk.